Allows the session to be reused for more than one task at a time.
This enables for instance to start an SSH connection, and then
use various channels over it.

The SSHChannel no longer is a subclass of an SSHSession, which made
things awkward when attempting to use the session again.

Instead, the session is a field in the channel, and new channels can
be constructed from the session if this is available.

This also turns off the sample that defaulted to publicKey
authentication.